Clean up solution
